fix: migrated ToastDialog into Toast component
